+++ starfish/network/convex/convex_account.py
@@ -1,0 +1,11 @@
+"""
+
+    Account class to provide basic functionality for convex network
+
+"""
+
+from starfish.network.account_base import AccountBase
+
+
+class ConvexAccount(AccountBase):
+    pass
