+++ starfish/network/convex/convex_network.py
@@ -1,0 +1,11 @@
+"""
+
+    Network class to provide basic functionality for convex network
+
+"""
+
+from starfish.network.network_base import NetworkBase
+
+
+class ConvexNetwork(NetworkBase):
+    pass
