+++ starfish/network/ethereum/contract/did_registry_contract.py
@@ -7,7 +7,7 @@
 import logging
 
 from starfish.network.ethereum.contract.contract_base import ContractBase
-from starfish.network.ethereum_account import EthereumAccount
+from starfish.network.ethereum.ethereum_account import EthereumAccount
 from starfish.utils.did import did_to_id
 
 logger = logging.getLogger(__name__)
