+++ starfish/network/ethereum/contract/direct_purchase_contract.py
@@ -4,7 +4,7 @@
 
 """
 
-from starfish.network.ethereum_account import EthereumAccount
+from starfish.network.ethereum.ethereum_account import EthereumAccount
 from starfish.types import AccountAddress
 
 from .contract_base import ContractBase
