+++ starfish/tool/command/account_create_command.py
@@ -7,7 +7,7 @@
 import logging
 from typing import Any
 
-from starfish.network.ethereum_account import EthereumAccount
+from starfish.network.ethereum.ethereum_account import EthereumAccount
 from .command_base import CommandBase
 
 logger = logging.getLogger(__name__)
