+++ starfish/tool/command/account_get_ether_command.py
@@ -6,7 +6,7 @@
 from typing import Any
 from web3 import Web3
 
-from starfish.network.ethereum_account import EthereumAccount
+from starfish.network.ethereum.ethereum_account import EthereumAccount
 
 from .command_base import CommandBase
