--- tests/resources/account_key_files/nile/key_file_f34d557105905ea7c6cf0941a551103174defac7.json
@@ -1,22 +1,0 @@
-{
-  "address": "f34d557105905ea7c6cf0941a551103174defac7",
-  "crypto": {
-    "cipher": "aes-128-ctr",
-    "cipherparams": {
-      "iv": "531892021387fc8d9ffdc484cc77e0c9"
-    },
-    "ciphertext": "ab4fa6ad8db224a3ed3808f21c1b3376f4996332fe3e4a6f23ae94ce071a749d",
-    "kdf": "scrypt",
-    "kdfparams": {
-      "dklen": 32,
-      "n": 262144,
-      "p": 8,
-      "r": 1,
-      "salt": "91cac6e0fc1cee603b96159a26564a40"
-    },
-    "mac": "821c847355d48df46a5392f1af4b5b5d87859e27f6c97a0f4aad401d1ece946f"
-  },
-  "id": "b6b63fbe-e914-449a-9600-c224311e3066",
-  "version": 3
-}
-
