+++ tests/unit/account/test_ethereum_account.py
@@ -7,7 +7,7 @@ import os
 
 from web3 import Web3
 
-from starfish.network.ethereum_account import EthereumAccount
+from starfish.network.ethereum.ethereum_account import EthereumAccount
 
 
 def test_init(config):
