+++ tests/unit/libs/unit_test_network.py
@@ -1,5 +1,5 @@
 
-from starfish.network.ethereum_network import EthereumNetwork
+from starfish.network.ethereum.ethereum_network import EthereumNetwork
 
 class UnitTestNetwork(EthereumNetwork):
